requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-multipart
itsdangerous
playwright
openai
python-dotenv
